# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProgramInfoNetwork =
  (typeof ProgramInfoNetwork)[keyof typeof ProgramInfoNetwork];

export const ProgramInfoNetwork = {
  devnet: "devnet",
  "mainnet-beta": "mainnet-beta",
} as const;

export type CommitmentStoreState = "ready" | "stuck" | "clean";

export interface ProgramInfo {
  /** On-chain program public key */
  programId: string;
  network: ProgramInfoNetwork;
  rpcEndpoint: string;
  /** Public key of the relayer keypair (for fee estimation) */
  relayerPublicKey: string;
  /** On-chain state of the commitment store PDA */
  commitmentStoreState: CommitmentStoreState;
}

export interface RelayWithdrawRequest {
  /** 64-char hex string (32 bytes). Relayer computes commitment=sha256(secret) to find the on-chain entry. */
  secret: string;
  /** Base58 public key of recipient wallet */
  recipient: string;
}

export interface RelayWithdrawResponse {
  /** Transaction signature on Solana */
  signature: string;
  /** Solana Explorer URL for the transaction */
  explorerUrl: string;
}

export interface TokenPrice {
  mint: string;
  symbol?: string;
  usdPrice: number;
  /** 24-hour price change percentage */
  change24h?: number;
}

export interface RelayError {
  error: string;
  code?: string;
}

export type GetTokenPriceParams = {
  /**
   * Token mint address (use So11111111111111111111111111111111111111112 for SOL)
   */
  mint: string;
};