# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns the program ID, network, and RPC endpoint in use
 * @summary On-chain program metadata
 */
export const GetProgramInfoResponse = zod.object({
  programId: zod.string().describe("On-chain program public key"),
  network: zod.enum(["devnet", "mainnet-beta"]),
  rpcEndpoint: zod.string(),
  relayerPublicKey: zod
    .string()
    .describe("Public key of the relayer keypair (for fee estimation)"),
});

/**
 * Accepts a signed commitment-nullifier withdrawal request and submits it
on-chain using the relayer keypair so the recipient address never appears
in the deposit transaction.

 * @summary Relay a private withdrawal
 */
export const RelayWithdrawBody = zod.object({
  secret: zod
    .string()
    .describe(
      "64-char hex string (32 bytes). Relayer computes commitment=sha256(secret) to find the on-chain entry.",
    ),
  recipient: zod.string().describe("Base58 public key of recipient wallet"),
});

export const RelayWithdrawResponse = zod.object({
  signature: zod.string().describe("Transaction signature on Solana"),
  explorerUrl: zod.string().describe("Solana Explorer URL for the transaction"),
});

/**
 * Returns USD price for a token mint address via CoinGecko/Birdeye
 * @summary Token price lookup
 */
export const GetTokenPriceQueryParams = zod.object({
  mint: zod.coerce
    .string()
    .describe(
      "Token mint address (use So11111111111111111111111111111111111111112 for SOL)",
    ),
});

export const GetTokenPriceResponse = zod.object({
  mint: zod.string(),
  symbol: zod.string().optional(),
  usdPrice: zod.number(),
  change24h: zod
    .number()
    .optional()
    .describe("24-hour price change percentage"),
});